+++ tests/conftest.py
@@ -12,6 +12,7 @@ https://docs.pytest.org/latest/monkeypatch.html
 """
 
 from pytest import fixture
+from transformers import BertModel
 
 from app.utils.configure_logging import toggle_global_debug_state
 
@@ -21,8 +22,37 @@ if True:
 from app.utils.load_configs import get_defaults, load_defaults_into_load_configs_module
 
 
-@fixture(scope="function", autouse=True)
+@fixture(scope="function")
 def save_dir_fixture() -> str:
-    """This is a test fixture with scope 'class' which returns int(1)"""
+    """Returns `save_dir` from the defaults configuration"""
+
     load_defaults_into_load_configs_module()
     return get_defaults("save_dir")
+
+
+@fixture(scope="function", params=["accuracy", "recall"])
+def metrics_to_test_for(request) -> list[str]:
+    """Returns strings of metric names"""
+    return request.param
+
+
+# "distilbert-base-uncased",
+# "google/electra-small-discriminator",
+# "microsoft/deberta-base",
+@fixture(scope="function", params=["bert-base-uncased"])
+def model_full_names(request):
+    """Returns strings of model names"""
+    return request.param
+
+
+@fixture(scope="function", params=[2])
+def num_labels(request):
+    """Returns ints of numbers of labels"""
+    return request.param
+
+
+# , DistilBertModel, ElectraModel, DebertaModel
+@fixture(scope="function", params=[BertModel])
+def subclasses_expected(request):
+    """Returns ints of numbers of labels"""
+    return request.param