+++ tests/test_load_hf_components.py
@@ -1,12 +1,14 @@
 #!/usr/bin/env python
-"""Test the load_hf_components"""
+# parametrized fixtures with metafunc
+# https://medium.com/opsops/deepdive-into-pytest-parametrization-cb21665c05b9
+# https://pytest.org/en/7.1.x/example/parametrize.html#deferring-the-setup-of-parametrized-resources
+# """Test the load_hf_components"""
 
 from logging import getLogger
 from os.path import join, sep
 
 from datasets.metric import Metric
 from pytest import mark
-from transformers import BertModel  # , DebertaModel, DistilBertModel, ElectraModel
 
 from app.utils.configure_logging import toggle_global_debug_state
 from app.utils.handle_paths import sanitize_path
@@ -16,7 +18,8 @@ if True:
 
 # delayed loading to set get_and_configure_logger:debug_on_global
 from app.utils.load_hf_components import (
-    get_metric_path_to_load,
+    get_list_of_metrics_to_load,
+    get_metric_path_or_name_to_load,
     get_model_hf,
     load_single_metric,
 )
@@ -24,26 +27,17 @@ from app.utils.load_hf_components import (
 logger = getLogger(__name__)
 
 
-@mark.usefixtures("save_dir_fixture")
-@mark.parametrize(
-    ["model_full_name", "num_labels", "subclass_expected"],
-    list(
-        zip(
-            [
-                "bert-base-uncased",
-                # "distilbert-base-uncased",
-                # "google/electra-small-discriminator",
-                # "microsoft/deberta-base",
-            ],
-            [2],  # , 2, 2, 2],
-            [BertModel],  # , DistilBertModel, ElectraModel, DebertaModel],
-        )
-    ),
+@mark.usefixtures(
+    "model_full_names", "num_labels", "subclasses_expected", "save_dir_fixture"
 )
-def test_get_model_hf(model_full_name, num_labels, subclass_expected, save_dir_fixture):
-    """TODO"""
-    model = get_model_hf(model_full_name, num_labels, save_dir_fixture)
-    assert isinstance(model, subclass_expected)
+def test_get_model_hf(
+    model_full_names, num_labels, subclasses_expected, save_dir_fixture
+):
+    """Expects models of specific subclasses"""
+
+    model = get_model_hf(model_full_names, num_labels, save_dir_fixture)
+
+    assert isinstance(model, subclasses_expected)
 
 
 # Objectives metrics loading
@@ -51,49 +45,40 @@ def test_get_model_hf(model_full_name, num_labels, subclass_expected, save_dir_f
 #   if not exists: from internet
 # 2) if internet: move metric from local cache to save_dir
 # 3) return list[Metric]
-@mark.parametrize(
-    "metric_name", ["accuracy"]  # , "precision", "recall", "f1", "mse", "mae"]
-)
-def test_load_single_metric(metric_name):
-    """TODO"""
-    metric_loaded = load_single_metric(metric_name)
+@mark.usefixtures("metrics_to_test_for")
+def test_load_single_metric(metrics_to_test_for):
+    """Expects a single valid Metric-object"""
+
+    metric_loaded = load_single_metric(metrics_to_test_for)
+
     assert isinstance(metric_loaded, Metric)
 
 
-@mark.usefixtures("save_dir_fixture")
-@mark.parametrize("metric_name", ["accuracy"])
-def test_get_metric_path_to_load(metric_name, save_dir_fixture):
-    """TODO"""
+@mark.usefixtures("metrics_to_test_for", "save_dir_fixture")
+def test_get_metric_path_or_name_to_load(metrics_to_test_for, save_dir_fixture):
+    """Expects a valid path to a Metric Builder Script"""
 
+    # TODO test for actual builder script, not only path
     dir = sanitize_path(save_dir_fixture)
     save_dir = join(dir["dir"], dir["base"])
-    expected_dir = f"{save_dir}{sep}Metrics{sep}{metric_name}"
+    expected_dir = f"{save_dir}{sep}Metrics{sep}{metrics_to_test_for}"
 
-    returned_dir = get_metric_path_to_load(metric_name, save_dir_fixture)
+    returned_dir = get_metric_path_or_name_to_load(
+        metrics_to_test_for, save_dir_fixture
+    )
 
     assert returned_dir == expected_dir
 
 
-# def test_get_metrics_to_load_objects_hf(metrics_to_load, save_dir):
-#     """TODO"""
-#     # Act
-#     model = get_metrics_to_load_objects_hf(metrics_to_load, save_dir)
-#     logger.debug("")
-#     logger.debug(f"{type(model)=}")
-#     # Assert
-#     assert type(model) == Metric
+@mark.usefixtures("metrics_to_test_for", "save_dir_fixture")
+def test_get_list_of_metrics_to_load(metrics_to_test_for, save_dir_fixture):
+    """Expects a list of valid Metric-objects"""
 
-# TODO parametrized fixtures
-# class TimeLine:
-#     def __init__(self, instances=[0, 0, 0]):
-#         self.instances = instances
+    metrics_returned = get_list_of_metrics_to_load(
+        [metrics_to_test_for], save_dir_fixture
+    )
 
-
-# @fixture(params=[[-2, 2, 30], [2, 4, 0], [6, 8, 10]])
-# def timeline(request):
-#     return TimeLine(request.param)
-
-
-# def test_timeline(timeline):
-#     for instance in timeline.instances:
-#         assert instance % 2 == 0
+    # TODO code smell multiple asserts in one test?
+    assert isinstance(metrics_returned, list)
+    for met in metrics_returned:
+        assert isinstance(met, Metric)